Refactoring source code directory
