+++ src/cli.js
